fix mailto dialog
closes #3972
