+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 - Fix crash on "Settings" click when not on main screen (e.g. no account selected): hide the "settings" button
 - code: comply with react hook rules #3955
+- fix mailto dialog
 
 <a id="1_46_1"></a>
