+++ scss/dialogs/_mailto-dialog.scss
@@ -1,4 +1,5 @@
 .mailto-dialog {
+  height: 100%;
   .select-chat-chat-list {
     height: 100%;
