+++ src/renderer/components/dialogs/MailtoDialog/styles.module.scss
@@ -1,0 +1,10 @@
+.MailtoDialog {
+  height: 100%;
+}
+
+.MailtoDialogBody {
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 0px;
+}
+
